update analysis results logic extraction
